Fix rename handling in git uploader when not run from repo root.
Fixes this bug:

Got error status from ['git', 'show',
'chromeos_login_manager/emit_login_prompt_ready.cc']:

"git show <pathname>" assumes a path relative to the current working
directory so it will error if not in the root of the repository.
Also, show will not actually show the file. To show the file, the
syntax is:

git show <revision>:<pathname>

With this syntax, the pathname is relative to the root of the repository.

Patch from Mandeep Singh Baines <msb@chromium.org>.