Add Quick LGTM and LGTM & CQ buttons
This permits reviewers to use their mouse to lgtm a change and increase the
likehood of CTS.

The Quick LGTM and CQ sets the commit bit.

Both are not present to the issue owner, the "CQ" flavor is not present if it is
already set or if the issue is closed.

Add a link to the CQ FAQ to act as a reference point.