Works around incorrect encoded files when sending email.
Replaces anything that is not valid utf-8 with strict ASCII.

Otherwise if the content is not encoded properly, the /publish endpoint will
fail because django.template.loader.render_to_string(template, context,
context_instance=RequestContext(request)) throws an DjangoUnicodeDecodeError.

Also add html meta header to specify the content is utf-8. It is already in the
HTTP header but adding it also to the html header doesn't hurt.

Review: http://codereview.appspot.com/6427052/