Move GetUserCredentials into a class so it can fetch the password from keyring only once.
When the password in the keyring is wrong
there is no way to enter a new one as each time around
the loop the wrong password is once again retrieved.
The solution is to only use the keyring for the first
attempt to login and fall back to asking the user
for the later ones.

BUG=329

Review URL: https://codereview.appspot.com/6492080/