Fix /publish's treatment of bare message_only POSTs one last time
Django form.initial does not work with bound forms. Initial is only meant for
initializing unbound forms for the purposes of display on an HTML output.  Once
a form is bound (as it's bound to POST in this case), initial is essentially
completely ignored.

This patch augments POST with the missing subject header, which sidesteps the
issue. If there is a current issue.subject, and the POST doesn't contain
subject, we then set it in the augmented post.

Author: iannucci@chromium.org

Review: https://codereview.appspot.com/7306089/