Work around incomplete patchset data.
This condition works around a case that reproduced on the chromium code review
tool. This logs an error instead of crashing but doesn't fix the problem.

The observed issue is:
Issue 11552033: 66001 is missing from {1L: 1, 8002L: 3, 8001L: 2, 30001L: 6,
22002L: 5, 19001L: 4, 75002L: 8, 42012L: 7}

It probably happened due to the AppEngine outage a month before.

Review: https://codereview.appspot.com/7314052/