Remove blocked user from issues.
Blocked users shouldn't participate in email communication anymore so
that they're not able to use the emails to send spam.

We've seen a blocked user that used the email communication to send
useless replies and to send private messages with spam to the
reviewers and ccs after a mail was sent out by the review server.

Review: http://codereview.appspot.com/10381045