Use issue and patchset in the try server job even if it is triggered.
Previously, for try server builds triggered by other builds, Rietveld would
attempt to recover issue and patchset numbers from the parent build. This proved
to be unreliable, since build numbers are not always unique.

This patch uses issue and patchset provided with the build, regardless of
whether the build is triggered or not, and only tries to guess when those values
are missing.

BUG: http://crbug.com/357677

On behalf of: Sergey Berezin

Code review URL: https://codereview.appspot.com/88170047/