Small fix to image serving for added or removed images.  When these images are missing, the ndb keys for them are missing too, so don't try to get().
The user visible affect is that users now see blank.jpg instead of a
broken image icon. And, more importantly, these requests do not add
500s to the server logs.

This was previously reviewed on the chromium branch but I forgot to
separate it out for the default branch.

Review URL: https://codereview.appspot.com/107490043/