Fixing SVN patch generation with new error code
The current error message in SVN when it can't find a remote file is `svn: E195012: Unable to find repository location for ...`.  However, the regex was looking for `svn: Unable to find repository location for ...` without the `E195012:`.  This patch allows the `E195012:` to optionally be present in the regex match.